separate out js logic into separate files
