+++ lib/school_house_web/templates/layout/_lesson_menu.html.leex
@@ -1,6 +1,6 @@
 <input type="checkbox" value="selected" id="toggle-one" class="toggle-input">
 <label for="toggle-one" class="block cursor-pointer py-6 px-4 lg:p-6 text-sm lg:text-base font-bold">Lessons</label>
-<div class="border-t-4 border-brand-purple-800 p-6 mega-menu md:absolute mb-0 md:mb-16 md:shadow-xl bg-nav dark:bg-nav-dark">
+<div class="border-t-4 border-brand-purple-800 pt-6 px-6 md:p-6 mega-menu md:absolute mb-6 md:mb-16 md:shadow-xl bg-nav dark:bg-nav-dark">
     <div class="container mx-auto w-full flex flex-wrap justify-left text-primary dark:text-primary-dark">
 
         <!-- First column -->
